Removed all shared drive mounting and finding functionality. Now all this module does is convert paths from local paths to platform agnostic ones and vice versa.
So if programs send paths over the network, they first call shared_drive.path_to_agnostic on it, and when programs receive paths, they call path_to_local before attempting to access the file.

This process is performed manually in many of our programs, they will slowly be migrated to use the same module, as is neater.