Fixed a rare bug where the first file in the list of files to watch was temporarily unavailable when getting the modified time but was available when checking if the path exists.
This bug may have occurred often, but if it was not the first file in the list of files, then it would have used the modified time for the previous file. No obvious error would have been reported, (file may have been shown to be modified) but it was still incorrect behaviour.