Applied an attempted workaround to what appears to be a python bug.
The modulewatcher.ModuleWatcher.check_and_unload function would crash sometimes if a garbage collection cycle was running at the same time
I am not sure which part of it specifically, as adding printlines seemed to decreas the probability of a collision.

The attempted workaround is to disable garbage collection and manually check once a cycle if garbage needs to be collected.