fixed bug in restoring modified times
we now check to see if files have been removed from the list of files to check, and remove those missing files from the list of modified times. This way if a files is removed from the list of files to check, and then readded, it won't show up as being modified if it happened to have been modified when we weren't monitoring it. I believe that this is the desired bahaviour!