Added camera_server.py, which provides a class CameraServer, for implementing BIAS-like camera servers in Python.
Subclassing it and overriding a few methods is sufficient to have some Python
code interact with a BLACS Camera tab. A basic example is included of a real
camera server I used to simply save images to file which were acquired with
an external program that didn't require any configuration info from labscript.

A Python camera server implementing this protocol via subclassing CameraServer
can be as simple or complex as required, perhaps just saving images to file as
in the minimalistic example provided, or being in sole control of cameras and
programming them with labscript-set parameters, performing fits, displaying
images, allowing you to set ROIs etc. This class just provides the interface to
BLACS - what else your camera server does with hdf5 shot files and cameras is up
to you.